requirements.txt
streamlit>=1.26.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
duckduckgo-search>=8.1.1
google-genai>=1.0.0
langchain>=0.0.171
